Enable the use of injectManifest in export
Previously it was commented out, this adds it back in so that the same
next configuration work for running both `next build` and `next export`
(otherwise it broke due to the arguments being different between
`InjectManifest` and `GenerateSW`).

The file contents of the `swSrc` parameter passed to `InjectManifest`
varies depending on if you're using the [webpack or direct node
implementation][1]. In webpack it looks for the
`self.__precacheManifest` variable where as the node version requires a
call to `workbox.precaching`. To enable both to work simultaneously
modify the `injectionPointRegexp` to look for the
`self.__precacheManifest` string which will definitely be there.

[1]: https://developers.google.com/web/tools/workbox/modules/workbox-webpack-plugin#injectManifest-swSrc